Move dhcp_conf.go from config to bootstrapper/dhcp
