+++ bootstrapper/dhcp/dhcp_conf.go
@@ -1,16 +1,16 @@
-package config
+package dhcp
 
 import (
 	"bytes"
 	"html/template"
-	"strings"
 
+	"github.com/k8sp/auto-install/config"
 	"github.com/topicai/candy"
 )
 
-// DHCPConf executes a template with a Cluster variable to generate
+// Conf executes a template with a Cluster variable to generate
 // /etc/dhcpd/dhcp.conf.
-func DHCPConf(tf string, c *Cluster) string {
+func Conf(tf string, c *config.Cluster) string {
 	tmpl := template.New("")
 
 	if len(tf) > 0 {
@@ -22,24 +22,6 @@ func DHCPConf(tf string, c *Cluster) string {
 	var buf bytes.Buffer
 	candy.Must(tmpl.Execute(&buf, c))
 	return buf.String()
-}
-
-// Join is defined as a method of Cluster, so can be called in
-// templates.  For more details, refer to const tmplDHCPConf.
-func (c Cluster) Join(s []string) string {
-	return strings.Join(s, ", ")
-}
-
-// Hostname is defined as a method of Node, so can be call in
-// template.  For more details, refer to const tmplDHCPConf.
-func (n Node) Hostname() string {
-	return strings.ToLower(strings.Replace(n.MAC, ":", "-", -1))
-}
-
-// Mac is defined as a method of Node, so can be called in template.
-// For more details, refer to const tmplDHCPConf.
-func (n Node) Mac() string {
-	return strings.ToLower(n.MAC)
 }
 
 const (
