+++ config/etcd_test.go
@@ -10,7 +10,7 @@ import (
 
 func TestInitialEtcdCluster(t *testing.T) {
 	c := &Cluster{}
-	candy.Must(yaml.Unmarshal([]byte(testConfig), c))
+	candy.Must(yaml.Unmarshal([]byte(ExampleYAML), c))
 	assert.Equal(t, c.InitialEtcdCluster(),
 		"00-25-90-c0-f7-80=http://10.10.10.201:2380,00-25-90-c0-f6-ee=http://10.10.10.202:2380,00-25-90-c0-f6-d6=http://00-25-90-c0-f6-d6:2380")
 }
